fix: allow sign when both language_icon & language_name are false
## Details

Added some early return logic when handling a code block language based
on the values of `language_icon` & `language_name`. However users might
still want signs in this case and that should be specifically disabled
via the `sign` option. It's unintuitive that disabling the other
`language_*` options results in disabling the `sign` as well.

The fix simply removes the early return. We'll still avoid adding
virtual text as it will end up empty later in the function.