feat: improve code language border for different conceal settings
## Details

When adding a thin border ignore whether the delimiter node is fully
concealed or not. Since we're doing an `overlay` it'll hide the
underlying text anyway, so whether it is concealed or not makes no
difference. Instead we just do a check that the delimiter node exists.

When computing leading spaces offset use the difference between the
number of leading spaces and the concealed width of the delimiter node.
This is a more generic version of the previous logic, which gated adding
the spaces behind the node being fully concealed.

Update unit test to handle more cases of code blocks.