feat: use builtin extends query in vim.treesitter.query.set for injections
## Details

With the neovim 0.11.0 release the vim.treesitter.query.set API was
updated to support the `extends` and `inherits` keywords much like
writing query files support but programmatically at runtime.

We can take advantage of this to simplify the logic of updating
`injections`, like we do for the `gitcommit` filetype if configured.

Rather than needing to loop through all the files append them to a
string, then add our injection, we can simply pre-pend the query with
the `;; extends` keyword and call set, no looping & appending needed.

Gate this new logic to only run if the user is on `0.11.0`, keep the old
logic around for the time being, eventually deprecate the previous
approach when we bump the minimum requirement for this plugin