feat: separate code border highlight, allow false value
## Details

Adds new `code.highlight_border` option which by default links to the
general `RenderMarkdownCode` highlight to avoid visible changes.

This value is used for both `thick` and `thin` borders as well as for
the background of the language name / icon section. The value is also
allowed to be a `boolean` (only a value of `false` makes sense but
`true` will have the same behavior as `false`). If set to a `boolean`
then no additional highlight is added for the language name / icon
section and bottom borders will have the same behavior as the `none`
value, i.e. they won't be rendered in any way.

Having `code.highlight_border = false` is the most equivalent to the
older behavior of `code.border = 'none'` which needed to be slightly
modified to accommodate more `border` values. Setting both of these is a
little redundant but up to the user.