feat: reduce height of LSP hover doc window based on concealed lines
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/384

Since the `conceal_lines` directive is enabled in the default highlights
neovim actively handles reducing the height of LSP hover doc windows to
remove unnecessary vertical space.

This happens here:

https://github.com/neovim/neovim/blob/master/runtime/lua/vim/lsp/util.lua#L1657-L1661

```lua
-- Reduce window height if TS highlighter conceals code block backticks.
local conceal_height = api.nvim_win_text_height(floating_winnr, {}).all
if conceal_height < api.nvim_win_get_height(floating_winnr) then
    api.nvim_win_set_height(floating_winnr, conceal_height)
end
```

Since our decorations are added separately from highlights and behave
very differently there would be no way for neovim to practically pickup
highlights added by this plugin.

To get around this we essentially re-implement this same logic as part
of this plugin. After the first load of a buffer when we are just about
to add our `extmark`s check if the buffer is related to LSP hover docs
and if it is adjust the height taking into account the concealed lines
that will be added (removed) by this plugin.

I don't see this as a bug, since adjusting windows to fit their content
is not something this plugin ever intended to do and supporting every
feature in neovim is not the goal. However, since this is for a very
specific scenario and the implementation is straightforward I've added
it.

> [!CAUTION]
> We'll likely need to update how we identify LSP hover doc buffers over
> time since the current approach is tied to internal undocumented behavior.

> [!NOTE]
> See if we can add a marker on LSP buffers that is a documented feature
> others can rely on, if such a thing doesn't already exist.