feat: wrap nvim_buf_set_extmark in pcall use notify_once if it errors
## Details

Related to issues like
https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/382

We use a version check like `vim.fn.has('nvim-0.11') == 1` before adding
any `extmarks` that use version specific options like `conceal_lines`.
By doing this we avoid ever getting to the API call that adds the marks
and can add fallback behavior for certain decorations like code borders.

However this version check has to make the assumption that the user has
all the features of the released version, which should be safe for most
people. However anyone using a dev version could have built it at a time
where the version check passes but before that feature was introduced.

This should be somewhat rare as it relies on someone having a dev build
and not upgrading when a release occurs.

In the event of an error we use `vim.notify_once` to inform the user
that they should update their build or switch to the released version
while not spamming errors at them.

Minor other changes:

- use `vim.o` to get options instead of `nvim_get_option_value`
- update `colorcolumn` suggestions to use option value