chore: move configs into individual modules
## Details

Rather than storing all type definition in `init.lua` & having all the
validation in `state.lua` create individual modules for each of the top
level configuration in a config folder and put the validation for these
in these modules.

Change the configuration naming pattern:

- from: `render.md.<module>` / `render.md.User<module>`
- to: `render.md.<module>.Config` / `render.md.<module>.UserConfig`

Makes writing code to automate the creation of user types easier and
more consistent. Users shouldn't really be relying on these but if they
are then they'll need to make changes.