feat: ignore option, checked before attaching
## Details

Adds top level `ignore` configuration option. This option should be a
function which accepts a buffer id as input and returns a boolean of
whether to ignore the buffer or not. Default value returns `false` for
all buffers to keep behavior unchanged.

This runs after the filetype & max file size checks. Users should use
this if they want to be more clever about the kinds of buffers this
plugin ignores, where the filetype, size, & option approaches are not
enough.

For instance you could use to check if a certain variable is set on the
buffer using `vim.api.nvim_buf_get_var` and ignore buffers that specify
certain variables.

At the time of attaching to a buffer it may not necessarily be
associated with a `window`, and even if it were this `window` can change
over time, so the function is not provided with a `window` id as part of
its input. Instead if users want to use the window of a `buffer` to
decide they'll need to do something like `local win = vim.fn.bufwinid(buf)`.

Minor other refactors:

- Update validator to not return self
- Name the local configs in renderers info