feat: add on.initial option called before adding marks to a buffer for the first time
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/396

Adds `on.initial` function which is called only once for a buffer before
marks are added on the first render cycle.

Context is provided with the buffer and window id. Modify other
callbacks called within the main UI loop to also take a window id,
meaning `on.render` & `on.clear` are both provided with `win`, since it
is a context table this breaks no existing usage.

This can be used to make changes to LSP hover docs based on the users
knowledge of how their plugins work but identifying buffers as LSP hover
docs and applying changes is left up to the user.