feat: better anti-conceal for multi-line decorations
## Details

Previously only the starting row of a mark would be used when deciding
whether the mark should be hidden or shown. This means if a mark spanned
multiple rows (like with scope highlights) we would only hide the
decoration if the users cursor was on the first line, but show it when
the cursor was anywhere in the middle.

This is okay, but definitely not ideal, so now the logic uses both the
start and end row of a mark and hides it if the current row overlaps the
mark. We use overlap since in visual mode it is not a single row, but
multiple. Now any overlapping marks within the cursor range are hidden.

There is a special case for `hl_eol` marks, which specify an end row but
do not actually impact that row, instead their impact effectively ends
one row earlier. So we handle this case specifically and offset the end
of a mark by 1 if `hl_eol` is specified. Hopefully there are no other
such edge cases.

Minor other changes:

- Add unit tests for `range` class
- Rename `render.md.Renderer` to `render.md.Render`
- Match order of fields in `default_buffer_config()` with class