chore: move find pattern ranges logic to node class
## Details

Finding the range of string patterns relative to a node really belongs
in the node class since it depends on the node's position along with its
text. Remove the old usage of `Str.find_all` + helper. Will make using
patterns in more places easier if we choose to do so.

Minor style change, inline `@type` annotations rather than having them
be on separate lines in a few contexts.

Update changelog.