chore: avoid calling tree:parse() when computing concealed ranges
## Details

I want to eventually take advantage of the async parsing capabilities
added to neovim. This gets easier if I limit where the `parse` function
is called to a single place.

To do this I just need to remove the only other place it's called which
is as part of calculating concealed regions. Honestly not sure why
that's there since we parse, iterate, then parse and iterate again?
Seems like the first call should do, though maybe I'm forgetting the
context behind this behavior. If it's a problem I can add a test for it
and add it back, but it seems safe enough.

Following this I'll experiment with moving some of the logic to
callbacks to make the transition easier.

Minor other changes:

- When logging runtime include a name and better align the output.
- For the benchmark tests do not include the amount of time it takes to
  run the callback, meaning feeding keys and kicking off an autocommand.
  How long this stuff takes is out of our control and leads to less
  consistent results. Update the thresholds based on this new approach.