feat: include removed values in config diff
## Details

The diff logic currently provides the user value if some part of the
configuration is modified. What this means is if a user provides fewer
values (by setting a list to the empty list for example) then no
difference is shown.

To fix this if the user value for some key is `nil` we set the value to
`vim.NIL`. That way we can identify values as having been removed
compared to the default in a unique way.

Minor other change:
- update ad-hoc unit tests to remove bullet point and add more link tests
- move setup of debug row logic out of api and into the module