feat: combine highlights for links
## Details

Request: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/413

Link icons are added using inline virtual text. Since we do not set an
`hl_mode` property the default value of `replace` is used. This is fine
in most context but in headings for instance, where we highlight the
background of the line the icons will continue to have their own default
background so will stick out.

The fix for this is to specify `hl_mode = combine` when we add link
icons, which results in the background getting picked up from the
background mark.

Minor other changes:

- Use method name `padding` consistently in renderers rather than
  different ones like `side_padding`, `left_pad`, etc.
- Add an enum for positions in node
- Raise an error when an invalid position is provided to `Node:line`,
  since this is only used internally it would mean there's definitely a
  bug somewhere that we'd otherwise swallow, include the row number in
  the output