feat: improve debug command
## Details

When debugging marks on current row we used the `config:hidden` method
which would return `nil` if anti conceal was disabled. Probably fairly
uncommon, but still, there is no need to rely on this.

So instead we create the `Range` for the current row directly when
debugging marks, bypassing any logic inside `config`.

This logic has also been almost entirely moved to the debug marks module
and out of ui. To support this ui publicly exposes what used to be in
`Cache.get(buf)` from which we can get `extmarks` using `get_marks()`.

Other minor changes:

- make `cache` part of each module rather than local variables
- move creation of main config into `Config` module
- replace `string.<method>(<s>, ...)` -> `(<s>):<method>(...)`
- rename `bg_to_fg` -> `bg_as_fg`
- rename `release_notification` -> `release`
- use local change of panvimdoc to pull `vim-version` dynamically
- update changelog