feat: use selene to prevent accidental print statements
## Details

After leaving a print statement in the last push, wanted an automated
way to prevent this in the future. While there's no perfect solution I
did come up with one that works well enough.

Added a `selene.toml` with a `neovim` base. The base defines the `vim`
global (as any) along with the various test framework globals like
`describe`, `it`, and the various `assert` methods. Important for this
change specifically it also marks both `print` and `vim.print` as
deprecated. That way any usage of those functions fails the check.

Where we do want to allow these methods add `selene: allow(deprecated)`.

It's not accurate since the methods are not deprecated, but does allow
me to quickly catch any usage that has not been explicitly allowed.

Added a `check` step that runs both `selene` & `stylua` over all files.

Fixed / improved a few issues:

- caught an error where I defined `Providers` globally instead of locally
- inline luassert `assert` methods so that errors are found
- add allow for mixed table around minimal config
- always provide a message argument to regular `assert` method