feat: use physical lines for table border if possible
## Details

Much like with heading borders, attempt to use physical lines if they
are empty, rather than always creating virtual lines. This is the new
default behavior, `pipe_table.border_virtual = true` can be set to
revert to the old behavior.

This logic does conflict with the indentation module, specifically in
that the definition of the range a specific section occupies is built
around headings and their borders, and tables would conflict with this,
leading to lines being at the incorrect indentation level.

Rather than trying something overly complicated where sections continue
to own the last empty line within themselves if that line belongs to a
pipe table, the `border_virtual` option is set to `true` if the indent
module is enabled, regardless of what the user has configured.

Minor other changes:

- When concealing text write functions that always take the full range
- Add a `used` module inside context to store which lines are occupied
  by any render for the request, replaces the `last_heading` field
- Use shorter `col` naming convention in table render rather than full
  `column` name
- Assert width of table cells >= 0 rather than handling these cases, it
  should be impossible