fix: don't discard events during debouncing
With the previous logic, if we were during a debouncing period and the
buffer was updated, the callback was never called.

This often happens with ai nvim plugins (e.g. codecompanion.nvim)
where the buffer is updated with a streaming response from an LLM.

After the response from the LLM was finished, the rendering of the
buffer was never updated until you either moved the cursor into the
buffer or switched modes.