feat: change code.border to thin instead of hide if version < 0.11.0
## Details

This does not change behavior as previously if we failed to hide the
border (which would happen on versions < 0.11.0), we would fallback to
using a thin border.

This change only makes this more visible (at the time of creating the
configuration), and users will be able to see this if they run the
`:RenderMarkdown config` command.

This allows us to remove the fallback logic in the rendering step,
instead we can now assume compatibility.