fix: correctly order indent & padding
## Details

Issue: https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/457

Indent extmarks and padding extmarks were both placed with priority 0,
this means if they were both put at the same row and column the order
they were rendered was non-deterministic. This would sometimes result in
the padding extmark going before the indent extmark, which is incorrect.

To fix this set the priority of padding to 100 and keep indent marks at
priority 0. This way the padding is added first, and then the indent,
which means the indent moves the padding over.