feat: change concealcursor value to nvic when user disables anti conceal
## Details

Have received a few bug reports when users disable the anti conceal
feature, then notice that while for some things it works, text that is
concealed in any way still becomes visible. For example with checkboxes
the original list marker becomes visible as shown here:
https://github.com/MeanderingProgrammer/render-markdown.nvim/issues/463

This is a reasonable expectation given the name of the feature but is a
misunderstanding of what is doing what. This plugin is adding
decorations using extmarks, when anti conceal is disabled we simply add
the marks on all the line, when it's enabled we avoid adding marks on
the current line. However conceal behavior on the current line is not
controlled by this plugin directly, and is instead controlled by the
concealcursor window option.

Previously for these issues I would inform users to update their
`win_options` configuration to set concealcursor to 'nvic'. At the same
time since this is behavior the user is going for we may as well make it
function that way out of the box for them, while still allowing them to
override the value for more custom experiences.

So the behavior now is if `anti_conceal.enabled == false`, the default
value for `win_options.concealcursor.rendered` becomes `'nvic'` instead
of `''`. This is handled similarly to a `preset`, even though that's not
what it is a `preset` is really just a way of setting different default
options based on the values of other options. If the user sets any value
in their config for `win_options.concealcursor.rendered` that will
always take priority and be used over the default.