feat: support quarto executable code block syntax
## Details

For `quarto`, executable code blocks are denoted by surrounding the
language name in curly braces. This is not valid `markdown` syntax
generally speaking but curly braces are handled by the parser, so this
plugin may as well support it as well, at least to some extent.

The way the parser is written leading spaces are ignored, and leading
curly braces are added to the info string, but not the language. Once
the language name ends (single word) any trailing text is also added to
the info string.

This change updates the logic of computing the info string around the
language to get any leading and trailing text, instead of assuming that
all text is trailing text. Text is positioned around the language in the
same way it would be in the underlying buffer. Info text can still be
disabled in the same way as before.